refactor: decompose critical hit resolution modules
